requirements.txt
asgiref==3.2.10
Django==3.1
django-debug-toolbar==2.2
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
coverage==5.3
requests==2.25.1
redis==3.5.3
django-redis
envparse==0.2.0
dj-database-url==0.5.0
python-dotenv==0.14.0
mysqlclient==2.0.1
